+++ src/utils/increaseOpacity.ts
@@ -1,6 +1,6 @@
 const increaseOpacity = (color) => {
   const [r, g, b, a] = color.match(/\d+/g).map(Number);
-  const increasedOpacity = Math.min(a + 0.2, 1);
+  const increasedOpacity = Math.min(a + 0.9, 1);
   return `rgba(${r}, ${g}, ${b}, ${increasedOpacity})`;
 };
